Use go/bqext v1.0.2
